### 2026-03-10 #289 @crhea93
dfcosmic: a Python package for cosmic ray removal
Submitting Author: (@crhea93 )
All current maintainers: (@crhea93 , @stevejanssens)
Package Name: dfcosmic
One-Line Description of Package: Python package for GPU accelerated cosmic ray removal following the original LACosmic algorithm
Repository Link:  https://github.com/DragonflyTelescope/dfcosmic
Version submitted: v0.0.1
Version accepted: TBD

---

- [x] I agree to abide by [pyOpenSci's Code of Conduct][PyOpenSciCodeOfConduct] during the review process and in future interactions in spaces supported by pyOpenSci should it be accepted.

`dfcosmic` is a high-performance computing implementation leveraging PyTorch for cosmic ray removal that strictly adheres to the original algorithm LACosmic developed by van Dokkum 2001. `dfcosmic` has significant computation gains over other python implementations of cosmic ray removal algorithms. Moreover, it is the only implementation that faithfully reproduces the results from the original IRAF code.

> [^1]: Please fill out a pre-submission inquiry before submitting a data visualization package.

The target audience is astronomers who need to remove cosmic rays from their data. This can be used in large scale pipelines and/or single observations.
There are a few very similar packages ([astroscrappy](https://github.com/astropy/astroscrappy) & [lacosmic](https://lacosmic.readthedocs.io/en/stable/)). However, they do not leverage CPU computing or CPU vectorization. More importantly, they depart in significant ways from the original algorithm which can mis-classify emission as cosmic rays.
N/A

Yes. I had already submitted this package to JOSS (https://github.com/openjournals/joss-reviews/issues/10161#issuecomment-4019970246). It was rejected on the basis of not having a public commit history of at least 6 months. I'd like to use this review in PyOpenSci to strengthen the codebase and eventually meet that criterion before re-submitting to JOSS.
