### 2026-03-12 bdae6a4
Memory safe chunking
