+++ src/dfcosmic/core.py
@@ -1,5 +1,6 @@
 import os
 from contextlib import nullcontext
+from time import perf_counter
 
 import numpy as np
 import torch
@@ -25,6 +26,25 @@ _KERNEL_CACHE: dict[
     tuple[str, torch.dtype],
     tuple[tuple[int, int], torch.Tensor, torch.Tensor, torch.Tensor],
 ] = {}
+
+
+def _current_rss_mb() -> float | None:
+    try:
+        with open("/proc/self/status", encoding="utf-8") as fh:
+            for line in fh:
+                if line.startswith("VmRSS:"):
+                    return float(line.split()[1]) / 1024.0
+    except Exception:
+        return None
+    return None
+
+
+def _log_rss(label: str) -> None:
+    rss_mb = _current_rss_mb()
+    if rss_mb is None:
+        print(f"[rss] {label}: unavailable")
+    else:
+        print(f"[rss] {label}: {rss_mb:.1f} MiB")
 
 
 def _get_kernels(device: torch.device, dtype: torch.dtype):
@@ -125,183 +145,239 @@ def lacosmic(
         if _THREADPOOLCTL_AVAILABLE:
             cpu_thread_ctx = threadpool_limits(limits=cpu_threads)
 
-    with cpu_thread_ctx:
-        # Move/cast to torch
-        if isinstance(image, np.ndarray):
-            image_t = torch.from_numpy(image).to(device).float().contiguous()
-        else:
-            image_t = image.to(device).float().contiguous()
+    prev_rss_env = os.environ.get("DFCOSMIC_RSS_DEBUG")
+    if rss_debug:
+        os.environ["DFCOSMIC_RSS_DEBUG"] = "1"
 
-        block_size_tuple, block_size_tensor, laplacian_kernel, strel = _get_kernels(
-            device, image_t.dtype
-        )
-        gkernel = torch.ones((3, 3), dtype=image_t.dtype, device=device)
+    try:
+        with cpu_thread_ctx:
+            if rss_debug:
+                _log_rss("start")
+            # Move/cast to torch
+            if isinstance(image, np.ndarray):
+                image_t = torch.from_numpy(image).to(device).float().contiguous()
+            else:
+                image_t = image.to(device).float().contiguous()
+            if rss_debug:
+                _log_rss("after input cast/to(device)")
 
-        clean_image = image_t.clone()
-        del image_t
-
-        final_crmask = torch.zeros(clean_image.shape, dtype=torch.bool, device=device)
-        if device.type == "cpu":
-            torch.backends.mkldnn.enabled = True
-            median_filter_fn = (
-                median_filter_cpp_torch if use_cpp_median else median_filter_torch
+            block_size_tuple, block_size_tensor, laplacian_kernel, strel = _get_kernels(
+                device, image_t.dtype
             )
-        else:
-            median_filter_fn = median_filter_torch
+            gkernel = torch.ones((3, 3), dtype=image_t.dtype, device=device)
 
-        with torch.no_grad():
-            for iteration in range(niter):
-                if verbose:
-                    print("")
-                    print(f"{'_' * 31} Iteration {iteration + 1} {'_' * 35}")
-                    print("")
+            clean_image = image_t.clone()
+            del image_t
+            if rss_debug:
+                _log_rss("after clean_image clone")
 
-                # Step 0: Gain estimation (if requested)
-                if gain <= 0:
-                    if verbose and iteration == 0:
-                        print("Trying to determine gain automatically:")
-                    elif verbose:
-                        print("Improving gain estimate:")
+            final_crmask = torch.zeros(
+                clean_image.shape, dtype=torch.bool, device=device
+            )
+            if rss_debug:
+                _log_rss("after final_crmask allocation")
[…]
