### 2026-03-12 682862b
chunk laplacian computation to reduce memory peak
