### 2026-03-12 54559f6
update ai disclosure
