### 2026-03-12 9c423a4
Update demos with new memory safe method
