### 2026-03-12 6093324
Force chunked conv for memory safety
