### 2026-03-12 1bf5f87
Add even more strict memory usage
