### 2026-03-16 #289 @kysolvik
Hi @crhea93!  Thank you for submitting your package for pyOpenSci

- [x] **Installation** The package can be installed from a community repository such as PyPI (preferred), and/or a community channel on conda (for example, conda-forge, bioconda).
- [x] **Documentation** The package has sufficient online documentation to allow us to evaluate the package's function and scope *without installing the package*. This includes:
  - [x] Short, quickstart tutorials that help a user understand how to use the package and what it can do for them.
  - [x] API documentation (documentation for your code's functions, classes, methods, and attributes): this includes clearly written docstrings with variables defined using a standard docstring format.
  - [x] **README** The package has a `README.md` file with a clear explanation of what the package does, instructions on how to install it, and a link to development instructions.
- [x] The package authors have disclosed the use of Generative AI tools in the development and/or maintenance of their package.

**Optional but recommended:**

- [x] **Linting** The package uses automated linting (e.g., [ruff](https://docs.astral.sh/ruff/)) to enforce code style and quality standards, ideally integrated with CI.
- [x] **Type checking** The package includes type hints and uses a type checker (e.g., [mypy](https://mypy-lang.org/)) to catch type-related issues, ideally integrated with CI.
- [ ] **Modern tooling** The package uses modern Python development tools (e.g., [uv](https://docs.astral.sh/uv/) for dependency management) and [Hatch](https://www.pyopensci.org/python-package-guide/package-structure-code/python-package-build-tools.html#about-hatch) for building publishing and running development tasks to streamline development workflows. See the pyOpenSci packaging guide for more on [modern packaging tools](https://www.pyopensci.org/python-package-guide/package-structure-code/python-package-build-tools.html).

---
A big thank you in advance to authors for setting aside five to ten minutes to do this. It truly helps our organization! :raised_hands:
---

Everything looks good @crhea93! I'll start the process of searching for an editor. A couple things in the meantime:

1. I checked the box for the quickstart tutorials, but it looks like the examples in the docs mostly focus on comparing to `astroscrappy` and `lacosmic`, which requires installing those extra packages. Maybe it'd be worth creating an even shorter quick-start that goes before those examples that focuses just on the barebones of applying cosmic ray removal from an example image (either synthetic or from the demos dir)?
2. Since the new submission template that includes this hasn't been finalized, could you briefly discuss if/how generative AI was used in the development of the package?
3. @stevejanssens could you fill out the [onboarding survey](https://forms.gle/F9mou7S3jhe8DMJ16) if you have the chance? This isn't required for anyone but the submitter, but it's great if all maintainers can fill this out.
