### 2026-03-18 ba1095e
Add quick usage example
