+++ paper.md
@@ -47,10 +47,10 @@ bibliography: dfcosmic.bib
 
 # Summary
 
-Astronomical images often show sharp features that are caused by cosmic ray (CR) hits, hot pixels, or non-Gaussian noise. L.A.Cosmic [@van_dokkum_cosmic-ray_2001] is a widely used edge detection algorithm that identifies and replaces such features. Here we describe `dfcosmic`, a direct python port of L.A.Cosmic utilizing PyTorch and C++ to enable efficient performance on both CPUs and GPUs. The code was developed for the MOTHRA array, which is projected to produce more than 1000 large format CMOS images every 15 minutes. Compared to previous pyhton implementations, `dfcosmic` achieves a speed gain of at least 20%.
+Astronomical images often show sharp features that are caused by cosmic ray (CR) hits, hot pixels, or non-Gaussian noise. L.A.Cosmic [@van_dokkum_cosmic-ray_2001] is a widely used edge detection algorithm that identifies and replaces such features. Here we describe `dfcosmic`, a direct python port of L.A.Cosmic utilizing PyTorch and C++ to enable efficient performance on both CPUs and GPUs. The code was developed for the MOTHRA array, which is projected to produce more than 1000 large format CMOS images every 15 minutes. Compared to previous Python implementations, `dfcosmic` achieves a speed gain of at least 20%.
 
 # Statement of need
-The Modular Optical Telephoto Hyperspectral Robotic Array (MOTHRA) uses CMOS sensors rather than traiditional CCds. Modern CMOS detectors have extraordinarily low noise but suffer from a relatively high number of hot pixels and non-Gaussian noise ("salt-and-pepper"; [@alarcon_scientific_2023]).
+The Modular Optical Telephoto Hyperspectral Robotic Array (MOTHRA) uses CMOS sensors rather than traditional CCDs. Modern CMOS detectors have extraordinarily low noise but suffer from a relatively high number of hot pixels and non-Gaussian noise ("salt-and-pepper"; [@alarcon_scientific_2023]).
 Therefore, the data reduction pipeline for MOTHRA requires rapid bright pixel identification and removal for tens of thousands of images every night using only a single core (2 threads) per frame.
 Although several implementations of L.A.Cosmic [@van_dokkum_cosmic-ray_2001] exist such as `lacosmic` [@bradley_larrybradleylacosmic_2025] and `astroscrappy` [@robitaille_astropyastroscrappy_2025], these implementations either deviate from the original algorithm in order to achieve computational gains or do not run fast enough for our usage. Importantly, experiments on the preliminary data taken by MOTHRA have demonstrated that it is crucial to use the original implementation (notably a true median filter rather than a separable median filter) in order to capture all the CRs (or hot pixels or salt-and-pepper non-Gaussian noise) without accidentally removing bright stars. `dfcosmic` has already been adopted in the nightly reduction pipeline for MOTHRA.
 
