### 2026-03-18 #289 @crhea93
So in the README I currently have the following short description of how generative AI was used:

Claude code was used to help with the unit tests only. CodeX was used to create only the c++ code for the median filter and make the code more memory efficient. Every line of code was manually inspected.
