### 2026-08-29 #289 @hamogu
@crhea93 Dear submitter,  I'm just writing here to apologize for the delay in getting this review off-the-ground 😿 .

I've contacted two reviewers, but both are unfortunately too busy at this time of year. In the meantime, the Editor in Chief has declared a moratorium on new reviews, until our editorial team is filled out with a sufficient number of editors again. While the terms of this moratorium seem to allow for a reasonable case-by-case judgement, I think it is appropriate to wait a little bit here until that situation settles. While we do that, I'm actively compiling a list of other potential reviewers I can reach out to.

Please be assured that I will do my best to not drag this out unnecessarily and I am prepared to continue slowly even while we have the general pause in place. Again, I apologize for the timing of this, and I suggest to address any question to @crhea93 using e.g. the PyOpenSci slack or reach out to me here.

[Yes, I know that the submitter is on the editorial team themselves. This message here is mostly for others who look over our issues to learn about the state of PyOpenSci.]
